% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm2utils-package.R
\docType{package}
\name{dsm2utils-package}
\alias{dsm2utils}
\alias{dsm2utils-package}
\title{dsm2utils: A Collection of Miscellaneous Utility Functions for DSM2 Tidefiles}
\description{
This package includes a variety of functions for working with DSM2 tidefiles. These include functions to obtain basic summary information about a tidefile, such as the start and end datetimes and the time step, and to extract channel flows for specified external channel numbers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dougjack.github.io/dsm2utils/}
}

}
\author{
\strong{Maintainer}: Doug Jackson \email{doug@QEDAconsulting.com}

}
\keyword{internal}
