% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFlows.R
\name{extractFlows}
\alias{extractFlows}
\title{Extract flows and stages for specified external channel numbers}
\usage{
extractFlows(tidefiles, channelNums, outputDir, figWidth = 10, figHeight = 5)
}
\arguments{
\item{tidefiles}{List of full paths to tidefiles}

\item{channelNums}{List of external channel numbers}

\item{outputDir}{Full path to the output directory}

\item{figWidth}{Figure width in inches. Default = 10}

\item{figHeight}{Figure height in inches. Default = 5}
}
\value{
None
}
\description{
Extracts flows and stages for a list of tidefiles and external channel
numbers. The data are saved in *.csv files in the specified output directory.
Time series plots are also created.
}
\examples{
\dontrun{
extractFlows(c(system.file("extdata", "hist_v82.h5", package="dsm2utils", mustWork=TRUE)),
             c(7), ".")
}
}
