% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTransferFlows.R
\name{extractTransferFlows}
\alias{extractTransferFlows}
\title{Extract transfer flows and stages for specified transfer names}
\usage{
extractTransferFlows(tidefiles, transferNames, outputDir)
}
\arguments{
\item{tidefiles}{List of full paths to tidefiles}

\item{transferNames}{List of transfer names}

\item{outputDir}{Full path to the output directory}
}
\value{
None
}
\description{
Extracts transfer flows for a list of tidefiles and transfer names.
The data are saved in *.csv files in the specified output directory.
}
\examples{
\dontrun{
extractTransferFlows(c(system.file("extdata", "hist_v82.h5", package="dsm2utils", mustWork=TRUE)),
             c("if_337"), ".")
}
}
