% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTidefile.R
\name{summarizeTidefile}
\alias{summarizeTidefile}
\title{Provide basic information about the tidefile}
\usage{
summarizeTidefile(tidefile)
}
\arguments{
\item{tidefile}{Full path to the tidefile}
}
\value{
List containing basic information
}
\description{
Provide basic information about the tidefile
}
