% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTidefile.R
\name{summarizeTidefile}
\alias{summarizeTidefile}
\title{Provide basic information about the tidefile}
\usage{
summarizeTidefile(tidefile)
}
\arguments{
\item{tidefile}{Full path to the tidefile}
}
\value{
List containing basic information
}
\description{
Reads the tidefile's metadata and provides basic information including start datetime, end datetime,
and time step.
}
\examples{
out <- summarizeTidefile(system.file("extdata", "hist_v82.h5", package="dsm2utils", mustWork=TRUE))
}
